Create ObtainImageDetails use case class
The ObtainImageDetails class implements the ParameterizedUseCase and the
execute() function takes a String, representing the id of the photo from
which the details are wanted.

It also takes an ImageDetailsFetcher and an ImageDetailsReceiver in the
constructor. The receiver in this case takes an instance of
ImageDetails, containing the necessary information to show the picture on
the screen.