Create FetchPublicImages use case class
The FetchPublicImages use case class implements the ParameterizedUseCase,
which is a protocol for use case classes that take parameters in their
executions.

This use case takes two objects in its constructor: an PublicImageFetcher
and PublicImageReceiver protocol implementations. The first represents an
object that fetch images from source, and the latter represents an object
that will receive the images or deal with errors.

Two initial struct models were created: Query, representing the
parameters of the search, and ImageSummary, representing each result,
holding basic info of the images, but not the paths to the image sizes.