Create GetImageInfo use case class
The GetImageInfo use case class implements the ParameterizedUseCase,
which is a protocol for use case classes that take parameters in their
executions.

This use case takes two objects in its constructor: an ImageInfoFetcher
and ImageInfoReceiver protocol implementations. The first represents an
object that fetch the image info from source, and the latter represents
an object that will receive the info or deal with errors.

The fetcher takes a photo id and returns an instance of ImageInfo,
containing the title, description, taken date, view count, and
comments count.