Create simple Dependency Injection engine
Basically we have a DependencyProvider, which enables the registration
of types and their corresponding instances, as well as an instance(of:)
method, where you pass a type and it returns the instance.

The provider also supports the registration using a factory, or a
creation function for a particular type, which means that a new instance
is created for every call of instance(of:).