Implement an ImageInfoFetcher, hitting Flickr's API
The DefaultImageInfoFetcher class implements the ImageInfoFetcher
protocol and takes two arguments in the constructor: an HTTPClient and
an ImageInfoJSONParser.

It just uses the client to make the request, then passes the response to
the parser to get an instance of ImageInfo back.

An implementation of ImageInfoJSONParser was also created, called
DefaultImageInfoJSONParser.