Implement an ImageUrlsFetcher, hitting Flickr's API
The DefaultImageUrlsFetcher class implements the ImageUrlsFetcher
protocol and takes two arguments in the constructor: an HTTPClient and
an ImageUrlsJSONParser.

It just uses the client to make the request, then passes the response to
the parser to get an instance of ImageUrls back.

An implementation of ImageUrlsJSONParser was also created, called
DefaultImageUrlsJSONParser.