Implement a PublicImagesFetcher, hitting Flickr's API
The DefaultPublicImagesFetcher class implements the PublicImagesFetcher
protocol and takes two arguments in the constructor: an HTTPClient and
an ImageSummaryJSONParser.

It just uses the client to make the request, then passes the response to
the parser to get an array of ImageSummary back.

An implementation of ImageSummaryJSONParser was also created, called
DefaultImageSummaryJSONParser.